+++ inst/NEWS.md
@@ -1,10 +1,55 @@
 # rdecision 1.2.1.9001
 
-* Moved functions `gbp`, `gv2png` and `placeholder` from vutils.R (in vignettes)
-  to utils.R, and removed vutils.R. These functions are now available to the
-  package, and vignettes no longer need to refer to them in an external file.
-
-
+* Removed dependency on external R files from vignette SM02-TKR by direct
+  inclusion of model code from test-model-SM02.R. Test expectations replaced
+  by `stopifnot` assertions in the vignette to ensure that vignette model
+  results agree with published values. Removed test-model-SM02.R from
+  testthat folder. Replaced calls to pander and knitr::kable with base R table
+  output so that tables and plots are displayed when purled code is executed
+  using source with echo.
+* Removed dependency on external R files from vignette DT02-Tegaderm by direct
+  inclusion of model code from test-model-Tegaderm.R. Test expectations replaced
+  by `stopifnot` assertions in the vignette to ensure that vignette model
+  results agree with published values. Removed test-model-Tegaderm.R from
+  testthat folder and added extra tests to test-DecisionTree. Replaced calls to
+  pander and knitr::kable with base R table output so that tables and plots are
+  displayed when purled code is executed using source with echo.
+* Removed dependency on external R files from vignette SM01-HIV by direct
+  inclusion of model code from test-model-AZT.R. Test expectations replaced by
+  `stopifnot` assertions in the vignette to ensure that vignette model results
+  agree with published values. Removed test-model-AZT.R from testthat folder.
+  This is because some package build and check environments break during
+  vignette building if the vignette R markdown refers to external scripts.
+* Moved function `gbp` from vutils.R (in vignettes) to utils.R, and removed
+  vutils.R. This function is now available to the package, and vignettes no
+  longer need to refer to it in an external file. Removed functions
+  `gv2png` and `placeholder` which are no longer required due to use of
+  `igraph` for drawing models.
+* Removed suggested package dependency on `DiagrammeR` and added a suggested
+  package dependency on `igraph`. Diagrams of Markov models in package vignettes
+  and readme file are now drawn with package `igraph`. Extra code is required
+  to produce desired appearance, compared with using GraphViz, but the resultant
+  image files are compact and can be generated without the need for a command
+  line tool.
+* Removed argument `expressions` from `modvar_table` in classes `DecisionTree`
+  and `SemiMarkovModel`. This is because operands of expressions are included
+  in the table, and it makes no sense to exclude expression model variables
+  which depend on those operands.
+* Modified `modvars` function in various classes to use vectorization. Added
+  to base classes `Node` and `Edge` to avoid type checks.
+* Removed name attributes from internal representation of node and edge lists
+  provided directly or indirectly to class `Graph`. This was causing errors
+  under certain circumstances when creating graphs with named lists, e.g., 
+  `g <- Graph(V = list(a = nodea, b = nodeb), E = list(e = edgee))` because
+  internal graph logic tests using `identical` gave false negatives.  
+* Added methods `as_gml` to classes `Graph` and `Digraph` for exporting of
+  graphs in the Graph Modelling Language (GML). Graph representations written
+  in this format can be imported to other packages, such as `igraph` to
+  permit modification of nodes, edges and their attributes.
+* Added parameters `rankdir`, `width` and `height` to function `as_DOT` in
+  `Graph` for consistency with function of the same name in `Digraph`. Removed
+  requirement for all nodes to have labels before using them by identifying
+  nodes with their ID, and using the `label` attribute in the file.
 
 # rdecision 1.2.1